Move from Auth0 to AWS Cognito
Auth0 was OAuth2 provider for Ministry of Justrice.
AWS Cognito is being used by Cabinet Office, in particular the app
named Cola.

This change removes the library, flow and configuration for Auth0
setup and moves it over to Cola.

Instead of specifying domain for everybody to see, using
ALLOWED_DOMAINS an env var, which will hold the list of whitelisted
domians.

It's more configurable this way.